Do not scroll to top when clicking challenge tabs
f6b731d748c4433df9e1f40b298c936bb921a4b7 introduced a `href` attribute
on the anchors used as Bootstrap tabs.
As there is nothing in JS preventing the default click action, this
causes this page to scroll to top (sorry).

A simple fix would be to use `@onclick.prevent` with AlpineJS, but a
cleaner fix is to use button elements. Using anchors as tabs does not
really make sense here as users don't need to drag&drop the anchor.